# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Terrace Farming System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Farm {
  id: number;
  userId: number;
  farmerName: string;
  farmName: string;
  location: string;
  regionType: string;
  /** @nullable */
  terraceCount?: number | null;
  /** @nullable */
  farmSize?: string | null;
  /** @nullable */
  soilType?: string | null;
  /** @nullable */
  waterSource?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface FarmBody {
  farmerName: string;
  farmName: string;
  location: string;
  regionType: string;
  /** @nullable */
  terraceCount?: number | null;
  /** @nullable */
  farmSize?: string | null;
  /** @nullable */
  soilType?: string | null;
  /** @nullable */
  waterSource?: string | null;
}

export interface Crop {
  id: number;
  userId: number;
  /** @nullable */
  farmId?: number | null;
  cropName: string;
  cropType: string;
  /** @nullable */
  sowingDate?: string | null;
  growthStage: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CropBody {
  cropName: string;
  cropType: string;
  /** @nullable */
  sowingDate?: string | null;
  growthStage: string;
  /** @nullable */
  notes?: string | null;
}

export interface SensorReading {
  id: number;
  userId: number;
  /** @nullable */
  farmId?: number | null;
  /** @nullable */
  deviceId?: string | null;
  soilMoisture: number;
  temperature: number;
  humidity: number;
  lightIntensity: number;
  waterLevel: number;
  sourceType: string;
  createdAt: string;
}

export interface ReadingBody {
  soilMoisture: number;
  temperature: number;
  humidity: number;
  lightIntensity: number;
  waterLevel: number;
}

export interface DeviceReadingBody {
  deviceId: string;
  soilMoisture: number;
  temperature: number;
  humidity: number;
  lightIntensity: number;
  waterLevel: number;
}

export interface Alert {
  id: number;
  userId: number;
  /** @nullable */
  farmId?: number | null;
  title: string;
  message: string;
  severity: string;
  status: string;
  createdAt: string;
}

export interface Recommendation {
  id: number;
  userId: number;
  /** @nullable */
  farmId?: number | null;
  category: string;
  message: string;
  createdAt: string;
}

export interface Device {
  id: number;
  userId: number;
  /** @nullable */
  farmId?: number | null;
  deviceId: string;
  deviceName: string;
  deviceType: string;
  status: string;
  /** @nullable */
  lastSync?: string | null;
  createdAt: string;
}

export interface DeviceBody {
  deviceId: string;
  deviceName: string;
  deviceType: string;
}

export interface AnalyticsPoint {
  timestamp: string;
  soilMoisture: number;
  temperature: number;
  humidity: number;
  lightIntensity: number;
  waterLevel: number;
}

export interface AnalyticsData {
  points: AnalyticsPoint[];
  notes: string[];
}

export interface AnalyticsSummary {
  avgSoilMoisture: number;
  avgTemperature: number;
  avgHumidity: number;
  avgLightIntensity: number;
  avgWaterLevel: number;
  farmCondition: string;
  cropNotes: string[];
  totalReadings: number;
  activeAlerts: number;
}

export interface PriceEntry {
  id: number;
  userId: number;
  cropName: string;
  pricePerKg: number;
  marketName: string;
  unit: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreatePriceEntryBody {
  cropName: string;
  pricePerKg: number;
  marketName?: string;
  unit?: string;
  /** @nullable */
  notes?: string | null;
}

export interface PricePrediction {
  cropName: string;
  currentPrice: number;
  nextDayPrice: number;
  nextWeekPrice: number;
  trend: string;
  trendPercent: number;
  dataPoints: number;
  note: string;
}

export interface PriceSummaryItem {
  cropName: string;
  latestPrice: number;
  marketName: string;
  trend: string;
  trendPercent: number;
  entryCount: number;
}

export type GetReadingsParams = {
  limit?: number;
};

export type GetAlertsParams = {
  status?: string;
};

export type GetPricesParams = {
  cropName?: string;
  limit?: number;
};