# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Terrace Farming System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /farm:
    get:
      operationId: getFarm
      tags: [farm]
      summary: Get farm profile
      responses:
        "200":
          description: Farm profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "404":
          description: Farm not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createFarm
      tags: [farm]
      summary: Create farm profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmBody"
      responses:
        "201":
          description: Farm created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
    put:
      operationId: updateFarm
      tags: [farm]
      summary: Update farm profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmBody"
      responses:
        "200":
          description: Farm updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
  /crops:
    get:
      operationId: getCrops
      tags: [crops]
      summary: List all crops
      responses:
        "200":
          description: List of crops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Crop"
    post:
      operationId: createCrop
      tags: [crops]
      summary: Create a crop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CropBody"
      responses:
        "201":
          description: Crop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crop"
  /crops/{id}:
    put:
      operationId: updateCrop
      tags: [crops]
      summary: Update a crop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CropBody"
      responses:
        "200":
          description: Crop updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crop"
        "404":
          description: Crop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCrop
      tags: [crops]
      summary: Delete a crop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Crop deleted
        "404":
          description: Crop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    put:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /devices:
    get:
      operationId: getDevices
      tags: [devices]
      summary: List devices
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceBody"
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    put:
      operationId: updateDevice
      tags: [devices]
      summary: Update a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceBody"
      responses:
        "200":
          description: Device updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - createdAt
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    Farm:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        farmerName:
          type: string
        farmName:
          type: string
        location:
          type: string
        regionType:
          type: string
        terraceCount:
          type: ["integer", "null"]
        farmSize:
          type: ["string", "null"]
        soilType:
          type: ["string", "null"]
        waterSource:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - farmerName
        - farmName
        - location
        - regionType
        - createdAt
        - updatedAt
    FarmBody:
      type: object
      properties:
        farmerName:
          type: string
        farmName:
          type: string
        location:
          type: string
        regionType:
          type: string
        terraceCount:
          type: ["integer", "null"]
        farmSize:
          type: ["string", "null"]
        soilType:
          type: ["string", "null"]
        waterSource:
          type: ["string", "null"]
      required:
        - farmerName
        - farmName
        - location
        - regionType
    Crop:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        farmId:
          type: ["integer", "null"]
        cropName:
          type: string
        cropType:
          type: string
        sowingDate:
          type: ["string", "null"]
        growthStage:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - cropName
        - cropType
        - growthStage
        - createdAt
        - updatedAt
    CropBody:
      type: object
      properties:
        cropName:
          type: string
        cropType:
          type: string
        sowingDate:
          type: ["string", "null"]
        growthStage:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - cropName
        - cropType
        - growthStage
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        farmId:
          type: ["integer", "null"]
        title:
          type: string
        message:
          type: string
        severity:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - message
        - severity
        - status
        - createdAt
...
    Device:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        farmId:
          type: ["integer", "null"]
        deviceId:
          type: string
        deviceName:
          type: string
        deviceType:
          type: string
        status:
          type: string
        lastSync:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - deviceId
        - deviceName
        - deviceType
        - status
        - createdAt
    DeviceBody:
      type: object
      properties:
        deviceId:
          type: string
        deviceName:
          type: string
        deviceType:
          type: string
      required:
        - deviceId
        - deviceName
        - deviceType
...