# lib/api-zod/src/generated/types/cropBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Terrace Farming System API
 * OpenAPI spec version: 0.1.0
 */

export interface CropBody {
  cropName: string;
  cropType: string;
  /** @nullable */
  sowingDate?: string | null;
  growthStage: string;
  /** @nullable */
  notes?: string | null;
}